+++ snapshot-tests-jaxb-jakarta/src/main/java/de/skuzzle/test/snapshots/data/xml/XmlSnapshot.java
@@ -260,7 +260,7 @@ public final class XmlSnapshot implements StructuredDataProvider {
     }
 
     @FunctionalInterface
-    static interface MarshallerSupplier {
+    public static interface MarshallerSupplier {
 
         Marshaller createMarshaller(JAXBContext jaxbContext) throws JAXBException;
     }