+++ RELEASE_NOTES.md
@@ -15,6 +15,7 @@
 * [#64](https://github.com/skuzzle/snapshot-tests/issues/64): Migrate build to gradle
 * [#66](https://github.com/skuzzle/snapshot-tests/issues/66): XmlSnapshot.withEnableXPathDebugging: Boolean parameter 'enabled' not taken into account
 * [#71](https://github.com/skuzzle/snapshot-tests/issues/71): Gracefully support JUnit5 nested tests
+* [#73](https://github.com/skuzzle/snapshot-tests/issues/73): `MarshallerSupplier` functional interface is not public
 * Add recursive directory scanning via `DirectoriesFrom.recursive()`
 * Deprecate `PathFilter` in favor of `TestFileFilter` and `TestDirectoryFilter`
 * Improve format and information density of XPath debugging (can be enabled with `XmlSnapshot.withEnableXPathDebugging(true)`)